+++ src/data/chapters.ts
@@ -13,6 +13,7 @@ import { CHAPTER_2 } from './chapter-2';
 import { CHAPTER_20 } from './chapter-20';
 import { CHAPTER_21 } from './chapter-21';
 import { CHAPTER_22 } from './chapter-22';
+import { CHAPTER_23 } from './chapter-23';
 import { CHAPTER_3 } from './chapter-3';
 import { CHAPTER_4 } from './chapter-4';
 import { CHAPTER_5 } from './chapter-5';
@@ -294,4 +295,5 @@ export const CHAPTERS = [
   CHAPTER_20,
   CHAPTER_21,
   CHAPTER_22,
+  CHAPTER_23,
 ] as const;